src/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
Django==2.2.3
django-filter==2.2.0
django-jsonfield==1.2.0
djangorestframework==3.10.1
drfdocs==0.0.11
Faker==0.9.1
idna==2.8
importlib-metadata==0.19
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mixer==6.1.3
more-itertools==7.2.0
openapi-codec==1.3.2
packaging==19.1
pluggy==0.12.0
psycopg2-binary==2.8.3
py==1.8.0
Pygments==2.4.2
pyparsing==2.4.2
pytest==5.0.1
pytest-cov==2.7.1
pytest-django==3.5.1
python-dateutil==2.6.0
python-dotenv==0.10.3
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.5.2
